Service Access Security

+++ src/main/java/fr/coveat/app/controller/RestorerController.java
@@ -4,6 +4,7 @@ import fr.coveat.app.form.DishForm;
 import fr.coveat.app.model.Dish;
 import fr.coveat.app.model.Restaurant;
 import fr.coveat.app.repository.DishRepository;
+import fr.coveat.app.service.SecurityService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Controller;
@@ -12,30 +13,34 @@ import org.springframework.util.StringUtils;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
+import javax.servlet.http.HttpServletRequest;
 import java.io.IOException;
 import java.util.Base64;
 import java.util.Objects;
 
 @Controller
-public class RestorerController {
+public class RestorerController implements SecurityService {
     @Autowired
     private DishRepository dishRepository;
 
     @RequestMapping(value = {"/restorer/","/restorer"}, method = RequestMethod.GET )
-    public String getDish(Model model) {
+    public String getDish(Model model, HttpServletRequest request) {
+        if(!checkConnected(request, "restaurant")){return "redirect:/login_restorer";}
         model.addAttribute("dishes", dishRepository.findAll(Sort.by("id").descending()));
         return "restorer/dish_list";
     }
 
     @RequestMapping(value = {"/restorer/add_dish"}, method = RequestMethod.GET )
-    public String createDish(Model model) {
+    public String createDish(Model model, HttpServletRequest request) {
+        if(!checkConnected(request, "restaurant")){return "redirect:/login_restorer";}
 
         model.addAttribute("dishForm", new Dish());
         return "restorer/add_dish";
     }
 
     @RequestMapping(value = {"/restorer/add_dish"}, method = RequestMethod.POST )
-    public String postCreateDish(@ModelAttribute("dish") Dish dish, @RequestParam(value = "photo", required = false) MultipartFile multipartFile){
+    public String postCreateDish(@ModelAttribute("dish") Dish dish, @RequestParam(value = "photo", required = false) MultipartFile multipartFile, HttpServletRequest request) {
+        if(!checkConnected(request, "restaurant")){return "redirect:/login_restorer";}
         if(dish.getName() != null && dish.getPrice() != null && dish.getDescription() != null && dish.getImageUrl() != null){
             String name = dish.getName();
             Double price = dish.getPrice();
@@ -67,7 +72,8 @@ public class RestorerController {
     }
 
     @RequestMapping(value = {"/restorer/{id}/edit_dish"}, method = RequestMethod.GET )
-    public String editDish(@PathVariable("id") Long id, Model model) {
+    public String editDish(@PathVariable("id") Long id, Model model, HttpServletRequest request) {
+        if(!checkConnected(request, "restaurant")){return "redirect:/login_restorer";}
         if (dishRepository.existsById(id)) {
             model.addAttribute("dishForm", dishRepository.getOne(id));
             return "restorer/edit_dish";
@@ -76,7 +82,8 @@ public class RestorerController {
     }
 
     @RequestMapping(value = {"/restorer/{id}/edit_dish"}, method = RequestMethod.POST )
-    public String editDish(@PathVariable("id") Long id, @ModelAttribute("dishForm") DishForm dishForm, @RequestParam(value = "photo", required = false) MultipartFile multipartFile){
+    public String editDish(@PathVariable("id") Long id, @ModelAttribute("dishForm") DishForm dishForm, @RequestParam(value = "photo", required = false) MultipartFile multipartFile, HttpServletRequest request) {
+        if(!checkConnected(request, "restaurant")){return "redirect:/login_restorer";}
         if (dishRepository.existsById(id)) {
             Dish dish = dishRepository.getOne(id);
 
@@ -114,7 +121,8 @@ public class RestorerController {
     }
 
     @RequestMapping(value = {"/restorer/{id}/delete_dish"}, method = RequestMethod.GET )
-    public String deleteDish(@PathVariable("id") Long id) {
+    public String deleteDish(@PathVariable("id") Long id, HttpServletRequest request) {
+        if(!checkConnected(request, "restaurant")){return "redirect:/login_restorer";}
         if (dishRepository.existsById(id)) {
             dishRepository.deleteById(id);
         }
